refactor: missing imports, organisation, add indirect dependencies
