+++ packages/komodo_defi_sdk/example/pubspec.yaml
@@ -7,7 +7,9 @@ environment:
   sdk: ^3.7.0
 
 dependencies:
+  bloc: ^9.0.0
   decimal: ^3.2.1
+  equatable: ^2.0.7
   flutter:
     sdk: flutter
   flutter_bloc: ^9.0.0
